requirements.txt
lyricsgenius
transformers
torch
python-dotenv
datasets
accelerate
